Reading reading rig config from DB

+++ DataLib/BusConfigurationDB.cs
@@ -1,13 +1,14 @@
 ﻿
+using System;
 using BusMaster.Model;
 
 namespace CoreHambusCommonLibrary.DataLib
 {
   public class BusConfigurationDB
   {
-    public int? Id { get; set; }
+    public long? Id { get; set; }
     public string Name { get; set; } = "";
-    public int Version { get; set; }
+    public long Version { get; set; }
     public BusType BusType { get; set; } = BusType.Unknown;
 
     public string Configuration { get; set; } = "{}";
